# api/openapi.yaml
openapi: "3.0.3"
info:
  title: MLP Client API
  version: "0.1-draft-02"
  description: >
    The mailbox-owner and guest surface of an MLP domain, as served
    by the reference implementation (mlpd) under /api/v1. Wire
    conventions per the Client API companion document and D-170: one
    JSON dialect (snake_case, integers only, RFC 3339 UTC), errors
    as application/problem+json, mutations require the X-MLP-Client
    header (CSRF posture) and, on POST, accept a client-minted
    Idempotency-Key (D-169). The FEDERATION surface (/dispatch,
    /fulfill, /verdict, /ingest, /.well-known/medialet.json) is
    signature-driven and specified normatively by the MLP Core
    Specification, not here.
  license:
    name: Apache-2.0
servers:
  - url: /api/v1
security:
  - session: []
...
paths:
  /auth/password:
    post:
      tags: [auth]
      summary: Password-fallback login (D-161); issues a session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, password]
              properties:
                address: { type: string, example: novak@target.example }
                password: { type: string, format: password }
      responses:
        "200": { description: Session cookie set., content: { application/json: { schema: { $ref: "#/components/schemas/SessionCreated" } } } }
        "401": { $ref: "#/components/responses/Problem" }
...
  /sessions/{hash}:
    delete:
      tags: [auth]
      summary: Revoke one session.
      parameters: [ { name: hash, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Revoked. }
        "404": { $ref: "#/components/responses/Problem" }
...
  /o/{urn}/accept:
    post:
      tags: [media]
      summary: >
        Accept an offered object. Local possession completes
        instantly (offered→expected→available in one action —
        claims, resends, dedup; D-241); otherwise the deferred
        verdict upgrades at the origin and the transfer follows.
      parameters: [ { $ref: "#/components/parameters/urn" } ]
      responses:
        "200": { description: Accepted., content: { application/json: { schema: { $ref: "#/components/schemas/AcceptResult" } } } }
        "404": { $ref: "#/components/responses/Problem" }
...
  /objects/have:
    get:
      tags: [media]
      summary: Batch possession check for composer attach-by-reference.
      parameters: [ { name: urns, in: query, required: true, schema: { type: string }, description: comma-separated } ]
      responses:
        "200": { description: Possession map., content: { application/json: { schema: { type: object, additionalProperties: { type: boolean } } } } }
...
  /search:
    get:
      tags: [search]
      summary: Full-text search over this mailbox's messages — subjects, body derived text, media names, and text extracted from held media.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, maxLength: 256 }, description: "User query; sanitized into FTS terms (implicit AND, trailing * = prefix)." }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Grouped matches, newest first; snippets mark hits with square brackets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  query: { type: string }
                  match: { type: string, description: The sanitized FTS expression actually run. }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        message_id: { type: integer, format: int64 }
                        thread_id: { type: integer, format: int64 }
                        medialet_ca: { type: string }
                        received_at: { type: string, format: date-time }
                        subject: { type: string }
                        matches:
                          type: array
                          items:
                            type: object
                            properties:
                              via: { type: string, enum: [message, media] }
                              name: { type: string, description: Media name (via=media only). }
                              snippet: { type: string }
        "400": { description: Malformed q/limit/offset., content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  /settings:
    get:
      tags: [system]
      summary: Mailbox settings.
      responses:
        "200": { description: Settings map., content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    patch:
      tags: [system]
      summary: Merge settings.
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { "200": { description: Saved. } }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: mlp_session
      description: >
        Issued by password login, passkey login, or a guest claim.
        Every non-GET request additionally requires the X-MLP-Client
        header (any value — the CSRF posture, D-170); mutating POSTs
        accept Idempotency-Key (D-169).
  parameters:
...
    urn: { name: urn, in: path, required: true, schema: { $ref: "#/components/schemas/Urn" } }
...
  responses:
    Problem:
      description: An error, always application/problem+json.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Urn:
      type: string
      pattern: "^urn:mlet:b[a-z2-7]+$"
      description: BLAKE3-256 multihash, multibase base32-lower (D-25).
      example: urn:mlet:bdyqbcgucl4thqvg5tqzv66l4q6myvy3j647jojvokzpw2c4ohcctu6y
    Problem:
      type: object
      properties:
        type: { type: string, example: "urn:mlp:err:malformed" }
        title: { type: string, example: malformed }
        status: { type: integer }
        detail: { type: string }
    SessionCreated:
      type: object
      properties:
        mailbox_id: { type: integer, format: int64 }
...
    AcceptResult:
      type: object
      properties:
        urn: { $ref: "#/components/schemas/Urn" }
        state: { type: string, enum: [available, expected] }
        instant: { type: boolean, description: "True when local possession completed the accept on the spot (D-241)." }
...